fix gettimeofday, psfork, mutex..., done sum = 1

+++ philo/philolib.h
@@ -6,7 +6,7 @@
 /*   By: qdo <qdo@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/26 19:41:57 by qdo               #+#    #+#             */
-/*   Updated: 2024/04/27 12:30:35 by qdo              ###   ########.fr       */
+/*   Updated: 2024/04/27 14:04:51 by qdo              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,7 +48,7 @@ typedef struct s_philo
 
 t_fork_mutex	*ft_fork_fill(t_philo *philo);
 t_philo			*ft_philo_fill(int ac, char **av);
-int				ft_cnt_time_of_acti(t_philo *philo_i);
+long int		ft_cnt_time_of_acti(t_philo *philo_i);
 struct timeval	*ft_print_out(t_philo *philo_i, char *str);
 void			ft_philo_create(t_philo *philo, pthread_t *philo_id);
 void			ft_philojob_groupsum2(t_philo *philo);
